components/demos/instance/weaviate/src/requirements.txt
weaviate-client
python-dotenv
